requirements.txt
web3>=5.31.4,<6.0
requests>=2.28.1
datetime>=4.7
pathlib>=1.0.1
eth_abi>=2.0.0
setuptools>=65.5.1
tqdm>=4.64.1
pandas>=1.3.5
safe-eth-py>=5.1.0
eth-ape>=0.6.6
slack-sdk>=3.20.2
